Update Equipment Guide.md
I'll be clear, I'm against adding Gangut's gun to this guide. Obtaining the gun is only possible if you were an active player during Nothern Overture and managed to get 10k in points for a copy of Gangut, and therefore, her gun. Spares would only be possible if you got multiple copies of her gun as well. Not to mention how it has no applicability on any other ship aside from Gangut herself making classification in this guide a rather fruitless one.

To make a comparison, we have Jean Bart's Quadruple 380mm that she needs to perform with. However, the T3 variant could be farmed for (in past events) and now it can be obtained from the core shop. It has a niche (albeit dumb-fun) use in giving non-barraging BBs/BCs the shell count and screen coverage to be a surprisingly effective mobbing tool and qualifies it to be on the guide. 

Gangut's gun is pointless on anyone but her, hence my distaste for having it added to this guide, which serves as a reference for players to compare the guns they might have in their depot to figure out what purpose each one serves, and how well.

That being said, there's nothing inherently wrong with the descriptions you made. Perhaps it would serve its purpose in another guide?